dsa: edge cases found by stress-testing every file against brute force - one-node list in cycle II, INT_MIN/INT_MAX nodes in validate BST, a grid with no oranges, an empty meeting list, k=0 in contains-duplicate (base and REDO_1), each with a one-line guard; and in subarray product Arpan traced the k=1 case himself (an empty window keeps prod at 1, and 1 >= 1 walks past the array) and kept the i <= j stop with his own note, plus the [1], k=1 test

+++ 08_DSA/12_TREES/14_validate_bst.cpp
@@ -35,7 +35,7 @@ struct TreeNode
     TreeNode(int x) : val(x), left(nullptr), right(nullptr) {}
 };
 
-bool solve(TreeNode *root, int mini, int maxi)
+bool solve(TreeNode *root, long long mini, long long maxi)
 {
     if (!root)
     {
@@ -61,7 +61,7 @@ bool isValidBST(TreeNode *root)
 {
     if (!root)
         return true;
-    return solve(root, INT_MIN, INT_MAX);
+    return solve(root, LLONG_MIN, LLONG_MAX);
 }
 
 int main()
